store/wizard: extract system selectors into dedicated module
